+++ CHANGELOG.md
@@ -4,12 +4,15 @@
 
 ## History
 
-### develop: 0.13.x (2022 Xxx xx)
+### develop: 0.14.0-dev (2022 Xxx xx)
 
  - new: `Pencil::format` returns colour string without outputting it
  - new: `Pencil::input` reads terminal input
  - new: `Pencil` cache original text for reversing it in certain situations
  - new: `Colour::Default` system colour
+ - update: `Arguments` implements `\Inane\Stdlib\Converters\JSONable`
+ - update: `Arguments::toJSON` added $pretty option to format json
+ - update: `HelpScreen` fixes for *php 8.1*
  - fix: various little updates and fixes
 
 ### 0.13.0 (2022 Oct 10)